ctr: move pprof to commands package
