+++ cmd/ctr/commands/pprof/pprof_unix.go
@@ -1,6 +1,6 @@
 // +build !windows
 
-package main
+package pprof
 
 import "net"
